+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Added
 
-### Planned
-
 - Separate categories can be created
 - Multiple entries per category will be summed up
 - Categories can be flagged as income, saving or cost
+
+### Planned
+
 - A plan for weekly costs can be created
+- Overview of all income summed up and all financial expenses